requirements.txt
# requirements.txt

# Frameworks Web y Servidor
fastapi
uvicorn[standard]
streamlit

# Core AI & CrewAI
crewai==0.28.8
crewai-tools==0.1.7
langchain-openai==0.0.5
langchain-core==0.1.31
langchain-community==0.0.28
openai>=1.3.7,<2.0.0

# Pydantic (fijar versión para compatibilidad)
pydantic==2.6.1

# Persistencia Vectorial
chromadb

# Servicios Externos (Google Drive)
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# NUEVA Herramienta de Búsqueda
tavily-python

# Utilidades
python-dotenv
requests